Create Simple HTML Templete
This just contains boxes that we can place our methods inside and create some organization for the page. This is just a simple template that we can refer to as a start, nothing on the page is currently functional. The current image background in there does not show and will be improved on.