Add files via upload
This says its 43% but its actually around 93%. We cannot test the GUI and group component so its decrease the percentage for all. The GUI action listener shouldn't be come when I was testing but there is no way to remove it and it also decrease the overall percent. Its 93% for the classes I tested.